fix(planner): RuleDecision::index u16 -> u32 (audit-R7 #4)
The trace entry stored the rule index in `u16` and the populate site
performed an `index as u16` truncation cast on the natural `usize`
enumerate index. With the rule registry holding 24+ entries today and
no static cap on growth, the cast is silently lossy — a regression that
pushed the registry past 65k rules would surface as "all decisions
collapse to the same wrap-around index" rather than as a compile or
test failure.

Widen the field to `u32`, which fits any feasible rule count without
truncation risk, and update the populate site (`index as u32`) along
with the docstring noting the rationale.

+++ crates/tokenfs-algos/src/dispatch/planner/rule.rs
@@ -48,7 +48,12 @@ pub struct Rule {
 #[derive(Copy, Clone, Debug, Eq, PartialEq)]
 pub struct RuleDecision {
     /// Index of the rule in `RULES`.
-    pub index: u16,
+    ///
+    /// Widened from `u16` to `u32` to match the natural `usize`
+    /// enumerate index without truncation risk; the populate site
+    /// always fits in `u32` for any feasible rule count
+    /// (audit-R7 finding #4).
+    pub index: u32,
     /// Stable rule name.
     pub name: &'static str,
     /// True iff the rule's predicate matched.
